parquet_factory/requirements.txt
boto3==1.21.41
botocore==1.24.41
grpcio==1.44.0
grpcio-tools==1.44.0
jmespath==1.0.0
protobuf==3.20.0
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
